% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-absenteeism.R
\docType{data}
\name{absenteeism}
\alias{absenteeism}
\title{Absenteeism data}
\format{
A dataframe with 865 rows and 4 variables:
\describe{
\item{days_absent}{The number of unscheduled days of absence for the employee in the last calendar year}
\item{tenure}{The employee’s tenure at the company in years}
\item{is_manager}{A binary value indicating 1 if the employee is a manager and 0 if not}
\item{performance_rating}{The employee’s most recent performance score on an increasing scale from 1 to 5}
}
}
\usage{
absenteeism
}
\description{
Fictional data on absenteeism in a large technology company
}
\examples{
absenteeism
}
\keyword{datasets}
