% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-employee_performance.R
\docType{data}
\name{employee_performance}
\alias{employee_performance}
\title{Employee performance data}
\format{
A dataframe with 366 rows and 5 variables:
\describe{
\item{sales}{The annual sales of the individual in millions of dollars}
\item{new_customers}{The number of new customers acquired by the individual}
\item{region}{The region the individuals works in - North, South, East or West}
\item{gender}{The gender of the individual}
\item{rating}{The performance rating of the individual - 1 = Low, 2 = Middle, 3 = High}
}
}
\usage{
employee_performance
}
\description{
Fictional data on employee performance evaluation metrics for a group of salespeople.
}
\examples{
employee_performance
}
\keyword{datasets}
