% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-employee_survey.R
\docType{data}
\name{employee_survey}
\alias{employee_survey}
\title{Employee survey data}
\format{
A dataframe with 2833 rows and 14 variables:
\describe{
\item{Happiness}{The employee rating on their overall happiness}
\item{Ben1, Ben2, Ben3}{The employee rating on three questions related to employment benefits}
\item{Work1, Work2, Work3}{The employee rating on three questions related to general work environment}
\item{Man1, Man2, Man3}{The employee rating on three questions related to perceptions of management}
\item{Car1, Car2, Car3, Car4}{The employee rating on four questions related to perceptions of career prospects}
}
}
\usage{
employee_survey
}
\description{
Fictional data on the results of an engagement survey among company employees on a five-point Likert scale indicating increasingly positive sentiment
}
\examples{
employee_survey
}
\keyword{datasets}
