% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-graduates.R
\docType{data}
\name{graduates}
\alias{graduates}
\title{Graduate salary data}
\format{
A dataframe with 173 rows and 5 variables:
\describe{
\item{Major}{The specific subject major}
\item{Discipline}{The broad subject discipline}
\item{Total}{The number of graduates of working age in the US}
\item{Unemployment_rate}{The proportion of graduates currently unemployed}
\item{Median_salary}{The current median salary of those employed in US dollars}
}
}
\source{
\href{https://github.com/fivethirtyeight/data}{FiveThirtyEight}
}
\usage{
graduates
}
\description{
Data on graduate salaries in the United States
}
\examples{
graduates
}
\keyword{datasets}
