% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-learning.R
\docType{data}
\name{learning}
\alias{learning}
\title{Learning program feedback data}
\format{
A dataframe with 4974 rows and 8 variables:
\describe{
\item{idcode}{The unique ID code of the participant}
\item{rec}{A binary value indicating whether the participant would recommend the program to others}
\item{rel}{A rating from the participant on the relevance of the program to their work, where 1 is Very Low and 5 is Very High}
\item{fun}{A rating on how enjoyable and fun the participant found the program, where 1 is Very Low and 5 is Very High}
\item{clar}{A rating from the participant on the clarity of the content and teaching in the program, where 1 is Very Low and 5 is Very High}
\item{home}{A rating from the participant on the quality of the homework or project work in the program, where 1 is Very Low and 5 is Very High}
\item{class}{A rating from the participant on the quality of the overall class who attended the program, where 1 is Very Low and 5 is Very High}
\item{fac}{A rating from the participant on the quality of the program faculty and instructors, where 1 is Very Low and 5 is Very High}
}
}
\usage{
learning
}
\description{
Fictional data on feedback from participants in a set of learning programs
}
\examples{
learning
}
\keyword{datasets}
