% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-promotion.R
\docType{data}
\name{promotion}
\alias{promotion}
\title{Promotion data}
\format{
A dataframe with 1134 rows and 5 variables:
\describe{
\item{diverse}{A binary value indicating membership of a diversity group at the company}
\item{flexible}{A binary value indicating whether or not the individual worked part-time for at least 6 months}
\item{store}{A binary value indicating whether the individual joined in a position working in the retail stores}
\item{promoted}{A binary value indicating whether or not the individual was promoted}
\item{result}{The year of the last record of the individual, where the date they joined was year 0.  If the individual was promoted, this will be the year of the promotion.}
}
}
\usage{
promotion
}
\description{
Fictional data on promotions in a retail company
}
\examples{
promotion
}
\keyword{datasets}
