% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-salespeople.R
\docType{data}
\name{salespeople}
\alias{salespeople}
\title{Salespeople promotion data}
\format{
A dataframe with 351 rows and 4 variables:
\describe{
\item{promoted}{A binary value indicating 1 if the individual was promoted and 0 if not}
\item{sales}{The sales (in thousands of dollars) attributed to the individual in the period of the promotion}
\item{customer_rate}{The average satisfaction rating from a survey of the individual’s customers during the promotion period}
\item{performance}{The most recent performance rating prior to promotion from 1 (lowest) to 4 (highest)}
}
}
\usage{
salespeople
}
\description{
Fictional data on promotion and performance for salespeople in a technology company
}
\examples{
salespeople
}
\keyword{datasets}
