% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sociological_data.R
\docType{data}
\name{sociological_data}
\alias{sociological_data}
\title{Sociological survey data}
\format{
A dataframe with 2618 rows and 9 variables:
\describe{
\item{annual_income_ppp}{The annual income of the individual in PPP adjusted US dollars}
\item{average_wk_hrs}{The average number of hours per week worked by the individual}
\item{education_months}{The total number of months spend by the individual in formal primary, secondary and tertiary education}
\item{region}{The region of the world where the individual lives}
\item{job_type}{Whether the individual works in a skilled or unskilled profession}
\item{gender}{The gender of the individual}
\item{family_size}{The size of the individual’s family of dependents}
\item{work_distance}{The distance between the indivdual’s residence and workplace in kilometers}
\item{languages}{The number of languages spoken fluently by the individual}
}
}
\usage{
sociological_data
}
\description{
Fictional data on a sociological survey related to income levels in various regions of the world.
}
\examples{
sociological_data
}
\keyword{datasets}
