% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ugtests.R
\docType{data}
\name{ugtests}
\alias{ugtests}
\title{Undergraduate examination data}
\format{
A dataframe with 975 rows and 4 variables:
\describe{
\item{Yr1}{Score in the first year examination on a scale of 0-100}
\item{Yr2}{Score in the second year examination on a scale of 0-200}
\item{Yr3}{Score in the third year examination on a scale of 0-200}
\item{Final}{Score in the final year examination on a scale of 0-300}
}
}
\usage{
ugtests
}
\description{
Fictional data on examination scores of undergraduates on a four year biology degree program.
}
\examples{
ugtests
}
\keyword{datasets}
