% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-health_insurance.R
\docType{data}
\name{health_insurance}
\alias{health_insurance}
\title{Health insurance data}
\format{
A dataframe with 1453 rows and 6 variables:
\describe{
\item{product}{The choice of policy the individual - A, B or C}
\item{age}{The age of the individual when they made the choice}
\item{children}{The number of child dependents the individual had at the time of the choice}
\item{position_level}{Position level in the company at the time they made the choice, where 1 is is the lowest and 5 is the highest}
\item{gender}{The gender of the individual as stated when they made the choice}
\item{tenure}{Full years employed by the company at the time they made the choice}
}
}
\usage{
health_insurance
}
\description{
Fictional data on the demographics and the choice of health insurance policy by employees of a large company
}
\examples{
health_insurance
}
\keyword{datasets}
